feat: moved command.ts to types dir
